+++ app/models/chat/app.js
@@ -32,9 +32,6 @@ class ChatApp extends AppCore {
     constructor(app) {
         super(app);
 
-        // Store then actived chat window id
-        this.activeChatWindow;
-
         // Handle application events
         app.on(R.event.app_socket_change, () => {
             this.socket.setHandler({
@@ -172,15 +169,29 @@ class ChatApp extends AppCore {
                         if(msg.isSuccess) {
                             if(this.user.isNewApi) {
                                 if(msg.data.gid) {
-                                    let chat = new Chat(msg.data);
+                                    let chat = this.dao.chats[msg.data.gid];
+                                    if(chat) {
+                                        chat.assign(msg.data);
+                                    } else {
+                                        chat = new Chat(msg.data);
+                                    }
                                     if(chat.isMember(this.user.id)) {
                                         chat.lastActiveTime = new Date().getTime();
                                         chat.updateWithApp(this);
                                         this.dao.updateChats(chat);
-                                    } else if(this.dao.chats[chat.gid]) {
+                                        if(chat.public && this.joinchatTask) {
+                                            setTimeout(() => {
+                                                this.$app.emit(R.event.ui_change, {
+                                                    navbar: R.ui.navbar_chat,
+                                                    activeChat: chat.gid
+                                                });
+                                            }, 500);
+                                        }
+                                    } else {
                                         this.dao.deleteChat(chat.gid);
                                     }
                                 }
+                                this.joinchatTask = false;
                             } else {
                                 if(!msg.data.join) {
                                     this.dao.deleteChat(msg.data.gid);
@@ -300,6 +311,23 @@ class ChatApp extends AppCore {
 
     get activeChat() {
         return this.activeChatWindow ? this.dao.getChat(this.activeChatWindow, true, true) : null;
+    }
+
+    get activeChatWindow() {
+        if(!this._activeChatWindow) {
+            let systemChats = this.dao.getChats(chat => {
+                return chat.isSystem;
+            });
+            if(systemChats && systemChats.length) {
+                this._activeChatWindow = systemChats[0].gid;
+            }
+        }
+        return this._activeChatWindow;
+    }
+
+    set activeChatWindow(chatGid) {
+        this.lastActiveChatWindow = this._activeChatWindow;
+        this._activeChatWindow = chatGid;
     }
 
     flashTrayIcon() {
@@ -1185,6 +1213,7 @@ class ChatApp extends AppCore {
      * @return {void}
      */
     joinChat(chat, join = true) {
+        this.joinchatTask = true;
         return this.socket.send(this.socket.createSocketMessage({
             'method': 'joinchat',
             'params': [chat.gid, join]